extract credits and quit panels
