+++ crates/client/src/ui/panels/mod.rs
@@ -1,3 +1,7 @@
+pub mod credits;
 pub mod help;
+pub mod quit;
 
+pub use credits::render_credits_panel;
 pub use help::render_help_panel;
+pub use quit::render_quit_panel;
